+++ temba/channels/types/dmark/type.py
@@ -20,10 +20,9 @@ class DMarkType(ChannelType):
     courier_url = r"^dk/(?P<uuid>[a-z0-9\-]+)/(?P<action>receive|status)$"
 
     claim_blurb = _(
-        """If you are based in Uganda or DRC you can purchase a short
-    code from <a href="http://dmarkmobile.com/">DMark Mobile</a> and connect it
-    in a few simple steps."""
-    )
+        "If you are based in Uganda or DRC you can purchase a short code from %(link)s and connect it in a few simple "
+        "steps."
+    ) % {"link": '<a href="http://dmarkmobile.com/">DMark Mobile</a>'}
     claim_view = ClaimView
 
     schemes = [URN.TEL_SCHEME]
@@ -31,9 +30,7 @@ class DMarkType(ChannelType):
     attachment_support = False
 
     configuration_blurb = _(
-        """
-        To finish configuring your DMark channel you need to set DMark to send MO messages to the URL below.
-        """
+        "To finish configuring your DMark channel you need to set DMark to send MO messages to the URL below."
     )
 
     configuration_urls = (