+++ temba/channels/types/junebug/type.py
@@ -18,18 +18,16 @@ class JunebugType(ChannelType):
     name = "Junebug"
     icon = "icon-junebug"
 
-    claim_blurb = _(
-        """Connect your <a href="https://junebug.praekelt.org/">Junebug</a> instance that you have already set up and configured."""
-    )
+    claim_blurb = _("Connect your %(link)s instance that you have already set up and configured.") % {
+        "link": '<a href="https://junebug.praekelt.org/">Junebug</a>'
+    }
     claim_view = ClaimView
 
     schemes = [URN.TEL_SCHEME]
     max_length = 1600
 
     configuration_blurb = _(
-        """
-        As a last step you'll need to configure Junebug to call the following URL for MO (incoming) messages.
-        """
+        "As a last step you'll need to configure Junebug to call the following URL for MO (incoming) messages."
     )
 
     configuration_urls = (
@@ -37,7 +35,8 @@ class JunebugType(ChannelType):
             label=_("Push Message URL"),
             url="https://{{ channel.callback_domain }}{% url 'courier.jn' channel.uuid 'inbound' %}",
             description=_(
-                "This endpoint will be called by Junebug when new messages are received to your number, it must be configured to be called as a POST"
+                "This endpoint will be called by Junebug when new messages are received to your number, it must be "
+                "configured to be called as a POST."
             ),
         ),
     )