+++ temba/channels/types/macrokiosk/type.py
@@ -18,9 +18,9 @@ class MacrokioskType(ChannelType):
 
     name = "Macrokiosk"
 
-    claim_blurb = _(
-        """Easily add a two way number you have configured with <a href="http://macrokiosk.com/">Macrokiosk</a> using their APIs."""
-    )
+    claim_blurb = _("Easily add a two way number you have configured with %(link)s using their APIs.") % {
+        "link": '<a href="http://macrokiosk.com/">Macrokiosk</a>'
+    }
     claim_view = ClaimView
 
     schemes = [URN.TEL_SCHEME]
@@ -29,9 +29,7 @@ class MacrokioskType(ChannelType):
     attachment_support = False
 
     configuration_blurb = _(
-        """
-        To finish configuring your MACROKIOSK connection you'll need to notify MACROKIOSK of the following URLs.
-        """
+        "To finish configuring your MACROKIOSK connection you'll need to notify MACROKIOSK of the following URLs."
     )
 
     configuration_urls = (