+++ temba/contacts/tests.py
@@ -4211,6 +4211,24 @@ class ContactFieldTest(TembaTest):
         )
 
         self.assertEqual(
+            ContactListView.prepare_sort_field_struct(sort_on="last_seen_on"),
+            (
+                "last_seen_on",
+                "asc",
+                {"field_type": "attribute", "sort_direction": "asc", "field_name": "last_seen_on"},
+            ),
+        )
+
+        self.assertEqual(
+            ContactListView.prepare_sort_field_struct(sort_on="-last_seen_on"),
+            (
+                "last_seen_on",
+                "desc",
+                {"field_type": "attribute", "sort_direction": "desc", "field_name": "last_seen_on"},
+            ),
+        )
+
+        self.assertEqual(
             ContactListView.prepare_sort_field_struct(sort_on="{}".format(str(self.contactfield_1.uuid))),
             (
                 str(self.contactfield_1.uuid),
@@ -5231,6 +5249,29 @@ class ESIntegrationTest(TembaNonAtomicTest):
             },
         )
 
+        now = timezone.now()
+        next_two_days = timezone.now() + timezone.timedelta(days=2)
+
+        self.create_contact(name="James", urns=["tel:+250188382999"], last_seen_on=next_two_days)
+        self.create_contact(name="Chris", urns=["tel:+250188382888"], last_seen_on=now)
+
+        # new contacts were created, execute the rp-indexer again
+        result = subprocess.run(
+            ["./rp-indexer", "-elastic-url", settings.ELASTICSEARCH_URL, "-db", database_url, "-rebuild"],
+            stdout=subprocess.PIPE,
+            stderr=subprocess.PIPE,
+        )
+        self.assertEqual(result.returncode, 0, "Command failed: %s\n\n%s" % (result.stdout, result.stderr))
+
+        # give ES some time to publish the results
+        time.sleep(5)
+
+        response = self.client.get("%s?sort_on=%s" % (url, "last_seen_on"))
+        self.assertEqual(response.context["object_list"][0].name, "Chris")  # oldest contact last seen
+
+        response = self.client.get("%s?sort_on=-%s" % (url, "last_seen_on"))
+        self.assertEqual(response.context["object_list"][0].name, "James")  # recent contact last seen
+
         # create a dynamic group on age
         self.login(self.admin)
         url = reverse("contacts.contactgroup_create")
@@ -5402,9 +5443,9 @@ class ContactImportTest(TembaTest):
         self.assertEqual(3, batches[0].record_end)
         self.assertEqual(
             [
-                {"name": "Eric Newcomer", "urns": ["tel:250788382382"], "groups": [str(imp.group.uuid)]},
-                {"name": "NIC POTTIER", "urns": ["tel:250(78) 8 383 383"], "groups": [str(imp.group.uuid)]},
-                {"name": "jen newcomer", "urns": ["tel:250788383385"], "groups": [str(imp.group.uuid)]},
+                {"name": "Eric Newcomer", "urns": ["tel:+250788382382"], "groups": [str(imp.group.uuid)]},
+                {"name": "NIC POTTIER", "urns": ["tel:+250788383383"], "groups": [str(imp.group.uuid)]},
+                {"name": "jen newcomer", "urns": ["tel:+250788383385"], "groups": [str(imp.group.uuid)]},
             ],
             batches[0].specs,
         )
@@ -5606,9 +5647,9 @@ class ContactImportTest(TembaTest):
 
         self.assertEqual(
             [
-                {"name": "Eric Newcomer", "urns": ["tel:250788382382"], "groups": [str(imp.group.uuid)]},
-                {"name": "NIC POTTIER", "urns": ["tel:250(78) 8 383 383"], "groups": [str(imp.group.uuid)]},
-                {"name": "jen newcomer", "urns": ["tel:250788383385"], "groups": [str(imp.group.uuid)]},
+                {"name": "Eric Newcomer", "urns": ["tel:+250788382382"], "groups": [str(imp.group.uuid)]},
+                {"name": "NIC POTTIER", "urns": ["tel:+250788383383"], "groups": [str(imp.group.uuid)]},
+                {"name": "jen newcomer", "urns": ["tel:+250788383385"], "groups": [str(imp.group.uuid)]},
             ],
             batch.specs,
         )